Return to regexp parsing
Might seem like bullshit. Will try to resolve in the future. Been
a while since I was on this project and by now this is beyond me.
Probably problem with using document, which should be avoided in
helpers. Cheerios solution or other plausibly appealing.

Know it's a step back. You must be disappointed. So am I. Need to
smoothen the shizz up after asciidoc update, before we start
refactoring again.